create useImageSize hook
